refactor(chat): extract touchAndCap* to shared collection-utils
